+++ sdk/postgresql/arm-postgresql-flexible/samples/v10-beta/javascript/tuningOptionsOperationsListRecommendationsSample.js
@@ -1,8 +1,8 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT License.
 
-import { PostgreSQLManagementFlexibleServerClient } from "@azure/arm-postgresql-flexible";
-import { DefaultAzureCredential } from "@azure/identity";
+const { PostgreSQLManagementFlexibleServerClient } = require("@azure/arm-postgresql-flexible");
+const { DefaultAzureCredential } = require("@azure/identity");
 
 /**
  * This sample demonstrates how to lists available object recommendations.
@@ -10,12 +10,12 @@ import { DefaultAzureCredential } from "@azure/identity";
  * @summary lists available object recommendations.
  * x-ms-original-file: 2026-04-01-preview/TuningOptionsListIndexRecommendations.json
  */
-async function listAvailableIndexRecommendations(): Promise<void> {
+async function listAvailableIndexRecommendations() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "ffffffff-ffff-ffff-ffff-ffffffffffff";
   const client = new PostgreSQLManagementFlexibleServerClient(credential, subscriptionId);
   const resArray = new Array();
-  for await (const item of client.tuningOptions.listRecommendations(
+  for await (const item of client.tuningOptionsOperations.listRecommendations(
     "exampleresourcegroup",
     "exampleserver",
     "index",
@@ -32,12 +32,12 @@ async function listAvailableIndexRecommendations(): Promise<void> {
  * @summary lists available object recommendations.
  * x-ms-original-file: 2026-04-01-preview/TuningOptionsListIndexRecommendationsFilteredForCreateIndex.json
  */
-async function listAvailableIndexRecommendationsFilteredToExclusivelyGetThoseOfCreateIndexType(): Promise<void> {
+async function listAvailableIndexRecommendationsFilteredToExclusivelyGetThoseOfCreateIndexType() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "ffffffff-ffff-ffff-ffff-ffffffffffff";
   const client = new PostgreSQLManagementFlexibleServerClient(credential, subscriptionId);
   const resArray = new Array();
-  for await (const item of client.tuningOptions.listRecommendations(
+  for await (const item of client.tuningOptionsOperations.listRecommendations(
     "exampleresourcegroup",
     "exampleserver",
     "index",
@@ -55,12 +55,12 @@ async function listAvailableIndexRecommendationsFilteredToExclusivelyGetThoseOfC
  * @summary lists available object recommendations.
  * x-ms-original-file: 2026-04-01-preview/TuningOptionsListTableRecommendations.json
  */
-async function listAvailableTableRecommendations(): Promise<void> {
+async function listAvailableTableRecommendations() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "ffffffff-ffff-ffff-ffff-ffffffffffff";
   const client = new PostgreSQLManagementFlexibleServerClient(credential, subscriptionId);
   const resArray = new Array();
-  for await (const item of client.tuningOptions.listRecommendations(
+  for await (const item of client.tuningOptionsOperations.listRecommendations(
     "exampleresourcegroup",
     "exampleserver",
     "table",
@@ -77,12 +77,12 @@ async function listAvailableTableRecommendations(): Promise<void> {
  * @summary lists available object recommendations.
  * x-ms-original-file: 2026-04-01-preview/TuningOptionsListTableRecommendationsFilteredForAnalyzeTable.json
  */
-async function listAvailableTableRecommendationsFilteredToExclusivelyGetThoseOfAnalyzeTableType(): Promise<void> {
+async function listAvailableTableRecommendationsFilteredToExclusivelyGetThoseOfAnalyzeTableType() {
   const credential = new DefaultAzureCredential();
   const subscriptionId = "ffffffff-ffff-ffff-ffff-ffffffffffff";
   const client = new PostgreSQLManagementFlexibleServerClient(credential, subscriptionId);
   const resArray = new Array();
-  for await (const item of client.tuningOptions.listRecommendations(
+  for await (const item of client.tuningOptionsOperations.listRecommendations(
     "exampleresourcegroup",
     "exampleserver",
     "table",
@@ -94,7 +94,7 @@ async function listAvailableTableRecommendationsFilteredToExclusivelyGetThoseOfA
   console.log(resArray);
 }
 
-async function main(): Promise<void> {
+async function main() {
   await listAvailableIndexRecommendations();
   await listAvailableIndexRecommendationsFilteredToExclusivelyGetThoseOfCreateIndexType();
   await listAvailableTableRecommendations();