+++ sdk/postgresql/arm-postgresql-flexible/src/classic/tuningOptionsOperations/index.ts
@@ -2,12 +2,16 @@
 // Licensed under the MIT License.
 
 import type { PostgreSQLManagementFlexibleServerContext } from "../../api/postgreSQLManagementFlexibleServerContext.js";
-import { listRecommendations, listByServer, get } from "../../api/tuningOptions/operations.js";
+import {
+  listRecommendations,
+  listByServer,
+  get,
+} from "../../api/tuningOptionsOperations/operations.js";
 import type {
-  TuningOptionsListRecommendationsOptionalParams,
-  TuningOptionsListByServerOptionalParams,
-  TuningOptionsGetOptionalParams,
-} from "../../api/tuningOptions/options.js";
+  TuningOptionsOperationsListRecommendationsOptionalParams,
+  TuningOptionsOperationsListByServerOptionalParams,
+  TuningOptionsOperationsGetOptionalParams,
+} from "../../api/tuningOptionsOperations/options.js";
 import type {
   TuningOptions,
   TuningOptionParameterEnum,
@@ -15,56 +19,56 @@ import type {
 } from "../../models/models.js";
 import type { PagedAsyncIterableIterator } from "../../static-helpers/pagingHelpers.js";
 
-/** Interface representing a TuningOptions operations. */
-export interface TuningOptionsOperations {
+/** Interface representing a TuningOptionsOperations operations. */
+export interface TuningOptionsOperationsOperations {
   /** Lists available object recommendations. */
   listRecommendations: (
     resourceGroupName: string,
     serverName: string,
     tuningOption: TuningOptionParameterEnum,
-    options?: TuningOptionsListRecommendationsOptionalParams,
+    options?: TuningOptionsOperationsListRecommendationsOptionalParams,
   ) => PagedAsyncIterableIterator<ObjectRecommendation>;
   /** Lists the tuning options of a server. */
   listByServer: (
     resourceGroupName: string,
     serverName: string,
-    options?: TuningOptionsListByServerOptionalParams,
+    options?: TuningOptionsOperationsListByServerOptionalParams,
   ) => PagedAsyncIterableIterator<TuningOptions>;
   /** Gets the tuning options of a server. */
   get: (
     resourceGroupName: string,
     serverName: string,
     tuningOption: TuningOptionParameterEnum,
-    options?: TuningOptionsGetOptionalParams,
+    options?: TuningOptionsOperationsGetOptionalParams,
   ) => Promise<TuningOptions>;
 }
 
-function _getTuningOptions(context: PostgreSQLManagementFlexibleServerContext) {
+function _getTuningOptionsOperations(context: PostgreSQLManagementFlexibleServerContext) {
   return {
     listRecommendations: (
       resourceGroupName: string,
       serverName: string,
       tuningOption: TuningOptionParameterEnum,
-      options?: TuningOptionsListRecommendationsOptionalParams,
+      options?: TuningOptionsOperationsListRecommendationsOptionalParams,
     ) => listRecommendations(context, resourceGroupName, serverName, tuningOption, options),
     listByServer: (
       resourceGroupName: string,
       serverName: string,
-      options?: TuningOptionsListByServerOptionalParams,
+      options?: TuningOptionsOperationsListByServerOptionalParams,
     ) => listByServer(context, resourceGroupName, serverName, options),
     get: (
       resourceGroupName: string,
       serverName: string,
       tuningOption: TuningOptionParameterEnum,
-      options?: TuningOptionsGetOptionalParams,
+      options?: TuningOptionsOperationsGetOptionalParams,
     ) => get(context, resourceGroupName, serverName, tuningOption, options),
   };
 }
 
-export function _getTuningOptionsOperations(
+export function _getTuningOptionsOperationsOperations(
   context: PostgreSQLManagementFlexibleServerContext,
-): TuningOptionsOperations {
+): TuningOptionsOperationsOperations {
   return {
-    ..._getTuningOptions(context),
+    ..._getTuningOptionsOperations(context),
   };
 }